+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -10,9 +10,9 @@ import device_info_plus
 import dynamic_color
 import network_info_plus_macos
 import package_info_plus
-import path_provider_macos
+import path_provider_foundation
 import screen_retriever
-import shared_preferences_macos
+import shared_preferences_foundation
 import url_launcher_macos
 import window_manager